# swagger.json
{
   "swagger":"2.0",
   "info":{
      "version":"1.0.0",
      "title":"Yet Another Node.js Blogg Application API",
      "description":"Yet Another Node.js Blogg Application API",
      "license":{
         "name":"MIT",
         "url":"https://opensource.org/licenses/MIT"
      }
   },
   "servers":[
      {
         "url":"/",
         "description":"Local Dev, or from Heroku"
      }
   ],
   "tags":[
      {
         "name":"Project",
         "description":"API for users in the system"
      }
   ],
   "schemes":[
      "http"
   ],
   "consumes":[
      "application/json"
   ],
   "produces":[
      "application/json"
   ],
   "paths":{
	   "/user/loginUser":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"Login User to application",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                      "$ref":"#/definitions/loginUser"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      },
		  "/user/registerUser":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To register User in application",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                      "$ref":"#/definitions/registerUser"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      },
		  "/user/getAllUser":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To get all userdetails",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                     
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      },
		  "/user/registerAdmin":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To register Admin in application",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                      "$ref":"#/definitions/registerAdmin"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      },"/user/loginAdmin":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"Login User to application",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                      "$ref":"#/definitions/loginAdmin"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	    }
		,"/expense/addExpense":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"Api to add expense",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                      "$ref":"#/definitions/addExpense"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
		  ,"/expense/fetchExpense":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"Api to add expense",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                      "$ref":"#/definitions/fetchExpense"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
		  ,"/property/addProperty":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To add Property",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                      "$ref":"#/definitions/addProperty"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
		  ,"/property/fetchProperty":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To fetch Property",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                     
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      },
		  "/booking/makeBooking":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To Add Booking",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                     "$ref":"#/definitions/makeBooking"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
		  ,
		  "/booking/fetchBookingByUser":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To fetch all Booking",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                     "$ref":"#/definitions/fetchBookingByUser"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
		   ,
		  "/booking/fetchBookingAll":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To fetch all Booking",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                     
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
		    ,
		  "/booking/fetchBookingDetails":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To fetch particular Booking details",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                     "$ref":"#/definitions/fetchBookingDetails"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
		  ,
		  "/booking/updateBooking":{
	         "post":{
	            "tags":[
	               "Project"
	            ],
	            "description":"To fetch particular Booking details",
	            "parameters":[
	               {
	                  "name":"fetch",
	                  "in":"body",
	                  "schema":{
	                     "$ref":"#/definitions/updateBooking"
	                  }
	               }
	            ],
	            "produces":[
	               "application/json"
	            ],
	            "responses":{
	               "200":{
	                  "description":"Ok",
	                  "schema":{
	                     
	                  }
	               }
	            }
	         }
	      }
   },
   "definitions":{
   "updateBooking":{
         "properties":{
			"bookingId":{
               "type":"integer"
            },
			"booking_on":{
               "type":"string"
            },
            "user_id":{
               "type":"integer"
            },
			"booking_by":{
               "type":"string"
            },
			 "booking_by_admin_id":{
               "type":"integer"
            },
			"booking_from":{
               "type":"string"
            },
			 "booking_to":{
               "type":"string"
            },
			"booking_rate":{
               "type":"integer"
            },
			 "booking_amount":{
               "type":"integer"
            },
			"property_id":{
               "type":"integer"
            },
			"add_on_services" : {
			"type": "array",
			
			"items":{"$ref": "#/definitions/expenseId"}
			
			}
         }
	},
   "fetchBookingDetails":{
         "properties":{
            "bookingId":{
               "type":"integer"
            }
         }
	},
   "fetchBookingByUser":{
         "properties":{
            "user_id":{
               "type":"string"
            }
         }
	},
	"makeBooking":{
         "properties":{
            "user_id":{
               "type":"integer"
            },
			"booking_by":{
               "type":"string"
            },
			 "booking_by_admin_id":{
               "type":"integer"
            },
			"booking_from":{
               "type":"string"
            },
			 "booking_to":{
               "type":"string"
            },
			"booking_rate":{
               "type":"integer"
            },
			 "booking_amount":{
               "type":"integer"
            },
			"property_id":{
               "type":"integer"
            },
			"add_on_services" : {
			"type": "array",
			
			"items":{"$ref": "#/definitions/expenseId"}
			
			}
         }
	},
	"addProperty":{
         "properties":{
            "property_name":{
               "type":"string"
            },
			"property_details":{
               "type":"string"
            }
         }
	},
   "fetchExpense":{
         "properties":{
            "room_type":{
               "type":"string"
            }
         }
	},
	"expenseId":{
         "properties":{
            "service_id":{
               "type":"integer"
            },
			 "amount":{
               "type":"integer"
            }
         }
	},
    "addExpense":{
         "properties":{
            "service_name":{
               "type":"string"
            },
            "room_type":{
               "type":"string"
            },
            "rate":{
               "type":"integer"
            }
         }
	},
    "loginUser":{
         "properties":{
            "type":{
               "type":"string"
            },
            "emailId":{
               "type":"string"
            },
            "socialId":{
               "type":"string"
            },
            "password":{
               "type":"string"
            }
         }
	},
	"loginAdmin":{
         "properties":{
           
            "emailId":{
               "type":"string"
            },
            "password":{
               "type":"string"
            }
         }
	},
	"registerUser":{
         "properties":{
			"firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            },
            "mob_no":{
               "type":"integer"
            },
            "emailId":{
               "type":"string"
            },
            "socialId":{
               "type":"string"
            },
            "password":{
               "type":"string"
            }
         }
	},
	"registerAdmin":{
         "properties":{
			"firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            },
            "mob_no":{
               "type":"integer"
            },
            "emailId":{
               "type":"string"
            },
            
            "password":{
               "type":"string"
            }
         }
	},
    "Response":{
         "properties":{
            "status":{
               "type":"string"
            },
            "message":{
               "type":"string"
            },
            "id":{
               "type":"string"
            }
          }
      }      
   }
}